Add data structures completely specifying the setup of the standard game
In particular, the connections within the map, the allowable variants
of the territory names, and the initial ownership and units and supply
centers.  This will allow, eventually, the removal of the Territory
and Subregion models/tables.

This information is not intended to change frequently, so it makes
more sense for it to be specified in a file, rather than in the
database.